requirements.txt
tensorflow
keras_tuner
streamlit
matplotlib
Pillow
numpy
pandas